style: UI 개선 및 개인정보처리방침 링크 추가
