+++ next.config.ts
@@ -10,6 +10,7 @@ const nextConfig = {
       "cdn.example.com",
       "image.yes24.com",
       "sprint-fe-project.s3.ap-northeast-2.amazonaws.com",
+      "lh3.googleusercontent.com",
     ],
   },
   eslint: {
