+++ src/app/auth/google/page.tsx
@@ -1,10 +1,11 @@
 "use client";
+import LoadingPage from "@/components/shared/Loading";
 import GoogleCallback from "@/features/google/components/GoogleCallback";
 import { Suspense } from "react";
 
 export default function GooglePage() {
   return (
-    <Suspense fallback={<p>구글 로그인 중입니다...⏳</p>}>
+    <Suspense fallback={<LoadingPage />}>
       <GoogleCallback />
     </Suspense>
   );
