+++ src/app/auth/kakao/page.tsx
@@ -1,10 +1,11 @@
 "use client";
 import { Suspense } from "react";
 import KakaoCallback from "../../../features/kakao/components/KakaoCallback";
+import LoadingPage from "@/components/shared/Loading";
 
 export default function KakaoPage() {
   return (
-    <Suspense fallback={<p>카카오 로그인 중입니다...⏳</p>}>
+    <Suspense fallback={<LoadingPage />}>
       <KakaoCallback />
     </Suspense>
   );
