+++ src/app/error.tsx
@@ -1,6 +1,8 @@
 "use client";
 import BaseButton from "@/components/shared/BaseButton";
 import { useRouter } from "next/navigation";
+import errorImg from "@/../public/animations/error.json";
+import Lottie from "lottie-react";
 
 export default function Error() {
   const router = useRouter();
@@ -12,13 +14,10 @@ export default function Error() {
   return (
     <div>
       <div className="flex flex-col items-center justify-center min-h-screen ">
-        <div className="relative">
-          <img
-            src="/error.svg"
-            alt="error Image"
-            className="md:w-[445px]  md:h-[249px] w-[326px] h-[182px]"
-          />
+        <div className="w-[200px] md:w-[250px]">
+          <Lottie animationData={errorImg} loop={true} />
         </div>
+        <p className="font-semibold text-3xl text-[var(--color-white)]">Oops! Something wrong.</p>
 
         <BaseButton
           className="lg:w-[440px] lg:h-[55px] md:w-[335px] md:h-[50px] w-[250px] h-[40px] lg:mt-[76px] md:mt-[36px] mt-[24px] md:text-[16px] text-[12px] font-extrabold"
