+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import LandingPage from "./landingpage/page";
 
-export default async function Home() {
+export default async function Home() {  
   return (
     <div>
       <main>
