+++ src/app/signup/error.tsx
@@ -1,5 +1,7 @@
 "use client";
 import BaseButton from "@/components/shared/BaseButton";
+import Lottie from "lottie-react";
+import abnormal from "@/../public/animations/Abnormal.json";
 import { useRouter } from "next/navigation";
 
 export default function ErrorPage() {
@@ -12,15 +14,13 @@ export default function ErrorPage() {
   return (
     <div>
       <div className="flex flex-col items-center justify-center min-h-screen ">
-        <img
-          src="/alarm.svg"
-          alt="error Image"
-          className="md:w-[166px]  md:h-[166px] w-[100px] h-[100px]"
-        />
-        <p className="font-semibold text-3xl text-[var(--color-white)] mt-[24px]">비정상적인 접근입니다.</p>
+        <div className="w-[200px] md:w-[250px]">
+          <Lottie animationData={abnormal} loop={true} />
+        </div>
+        <p className="font-semibold text-3xl text-[var(--color-white)] -mt-[34px]">비정상적인 접근입니다.</p>
 
         <BaseButton
-          className="lg:w-[440px] lg:h-[55px] md:w-[335px] md:h-[50px] w-[250px] h-[40px] lg:mt-[76px] md:mt-[36px] mt-[24px] md:text-[16px] text-[12px] font-extrabold"
+          className="lg:w-[440px] lg:h-[55px] md:w-[335px] md:h-[50px] w-[250px] h-[40px] lg:mt-[46px] md:mt-[16px] mt-[14px] md:text-[16px] text-[12px] font-extrabold"
           onClick={handleClick}
         >
           홈 화면으로 이동
